Reorganize API and object store messages
